requirements.txt
flask
psycopg2-binary
neonize
google-auth-oauthlib
google-api-python-client
groq
qrcode
pillow
gunicorn
requests
